Update schema, add db schema to version control
