+++ src/@types/custom.d.ts
@@ -2,11 +2,27 @@ import {PrismaClient} from "@prisma/client";
 import express, {Express} from "express";
 import http from "http";
 import {ApolloServer} from "apollo-server-express";
+import {Ability, RawRule} from "@casl/ability";
+
+declare module 'express-session' {
+    interface SessionData {
+        permissionJSON?: RawRule[];
+        userId?: number;
+    }
+}
+
+declare module 'express-serve-static-core' {
+    interface Request {
+        permissions?: Ability;
+        // user?: User;
+        prisma: PrismaClient;
+    }
+}
 
 type ResolverContext = {
-    prisma: PrismaClient,
-    req: Express.Request,
-    res: Express.Response
+    prisma: PrismaClient;
+    req: Express.Request;
+    res: Express.Response;
 }
 
 /**
